+++ app/src/main/java/org/breezyweather/remoteviews/presenters/HourlyTrendWidgetIMP.kt
@@ -80,7 +80,7 @@ object HourlyTrendWidgetIMP : AbstractRemoteViewsPresenter() {
     private fun getDrawableView(context: Context, location: Location?, cardStyle: String?): View? {
         val weather = location?.weather ?: return null
         val provider = ResourcesProviderFactory.newInstance
-        val itemCount = min(5, weather.next24HourlyForecast.size)
+        val itemCount = min(5, weather.nextHourlyForecast.size)
         var highestTemperature: Float? = null
         var lowestTemperature: Float? = null
         val minimalIcon = SettingsManager.getInstance(context).isWidgetUsingMonochromeIcons
@@ -96,7 +96,7 @@ object HourlyTrendWidgetIMP : AbstractRemoteViewsPresenter() {
         run {
             var i = 0
             while (i < temperatures.size) {
-                temperatures[i] = weather.next24HourlyForecast.getOrNull(i / 2)?.temperature?.temperature
+                temperatures[i] = weather.nextHourlyForecast.getOrNull(i / 2)?.temperature?.temperature
                 i += 2
             }
         }
@@ -116,7 +116,7 @@ object HourlyTrendWidgetIMP : AbstractRemoteViewsPresenter() {
             lowestTemperature = normals.nighttimeTemperature
         }
         for (i in 0 until itemCount) {
-            weather.next24HourlyForecast[i].temperature?.temperature?.let {
+            weather.nextHourlyForecast[i].temperature?.temperature?.let {
                 if (highestTemperature == null || it > highestTemperature!!) {
                     highestTemperature = it
                 }
@@ -156,7 +156,7 @@ object HourlyTrendWidgetIMP : AbstractRemoteViewsPresenter() {
             drawableView.findViewById(R.id.widget_trend_hourly_item_5)
         )
         widgetItemViews.forEachIndexed { i, widgetItemView ->
-            weather.next24HourlyForecast.getOrNull(i)?.let { hourly ->
+            weather.nextHourlyForecast.getOrNull(i)?.let { hourly ->
                 widgetItemView.setTitleText(hourly.getHour(context, location.timeZone))
                 widgetItemView.setSubtitleText(null)
                 hourly.weatherCode?.let {