+++ app/src/main/java/org/breezyweather/remoteviews/presenters/notification/NativeWidgetNotificationIMP.kt
@@ -20,6 +20,7 @@ import android.content.Context
 import android.graphics.drawable.Icon
 import android.os.Build
 import androidx.core.app.NotificationCompat
+import androidx.core.graphics.drawable.IconCompat
 import notificationBuilder
 import notify
 import org.breezyweather.R
@@ -79,11 +80,17 @@ object NativeWidgetNotificationIMP : AbstractRemoteViewsPresenter() {
         val notification = context.notificationBuilder(Notifications.CHANNEL_WIDGET).apply {
             priority = NotificationCompat.PRIORITY_MAX
             setVisibility(NotificationCompat.VISIBILITY_PUBLIC)
-            setSmallIcon(
-                if (temperature != null) {
-                    ResourceHelper.getTempIconId(context, temperatureUnit.getValueWithoutUnit(temperature).roundToInt())
-                } else ResourceHelper.getDefaultMinimalXmlIconId(current.weatherCode, daytime)
-            )
+            if (temperature != null && Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                setSmallIcon(
+                    IconCompat.createWithBitmap(
+                        ResourceHelper.createTempBitmap(context, temperature)
+                    )
+                )
+            } else {
+                setSmallIcon(
+                    ResourceHelper.getDefaultMinimalXmlIconId(current.weatherCode, daytime)
+                )
+            }
             if (current.weatherCode != null) setLargeIcon(
                 ResourceHelper.getWidgetNotificationIcon(
                     provider, current.weatherCode,